Data Validations added

+++ api/Controllers/StockController.cs
@@ -19,15 +19,21 @@ public class StockController : ControllerBase
     [HttpGet]
     public async Task<IActionResult> GetAll()
     {
+        if (!ModelState.IsValid)
+            return BadRequest(ModelState);
+        
         var stocks = await _stockRepository.GetAllAsync();
         var stocksDto = stocks.Select(s => s.ToStockDto());
 
         return Ok(stocksDto);
     }
 
-    [HttpGet("{id}")]
+    [HttpGet("{id:int}")]
     public async Task<IActionResult> GetById([FromRoute] int id)
     {
+        if (!ModelState.IsValid)
+            return BadRequest(ModelState);
+        
         var stock = await _stockRepository.GetByIdAsync(id);
 
         if (stock == null)
@@ -39,26 +45,35 @@ public class StockController : ControllerBase
     [HttpPost]
     public async Task<IActionResult> Create([FromBody] CreateStockRequestDto stockDto)
     {
+        if (!ModelState.IsValid)
+            return BadRequest(ModelState);
+        
         var stockModel = stockDto.ToStockFromCreateDto();
         await _stockRepository.CreateAsync(stockModel);
 
         return CreatedAtAction(nameof(GetById), new { id = stockModel.StockId }, stockModel.ToStockDto());
     }
 
-    [HttpPut("{id}")]
+    [HttpPut("{id:int}")]
     public async Task<IActionResult> Update([FromRoute] int id, [FromBody] UpdateStockRequestDto updateDto)
     {
+        if (!ModelState.IsValid)
+            return BadRequest(ModelState);
+        
         var updatedModel = await _stockRepository.UpdateAsync(id, updateDto);
         if (updatedModel == null)
             return BadRequest($"This stock with id {id} doesn't exists.");
 
         return Ok(updatedModel.ToStockDto()); 
     }  
-
+ 
     [HttpDelete]
-    [Route("{id}")]
+    [Route("{id:int}")]
     public async Task<IActionResult> Delete([FromRoute] int id)
     {
+        if (!ModelState.IsValid)
+            return BadRequest(ModelState);
+        
         var deletedStock = await _stockRepository.DeleteAsync(id);
 
         if (deletedStock == null)
